backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
flask-cors==4.0.0
Pillow==10.0.0
python-multipart==0.0.6
httpx==0.27.0